Registro de Clientes

+++ FrontEnd/CanchaLibre/CanchaLibreWeb/CanchaLibreWeb/ViewModels/ClienteViewModel.cs
@@ -4,6 +4,6 @@ namespace CanchaLibreWeb.ViewModels;
 
 public class ClienteViewModel : PersonaViewModel
 {
-    public int Calificacion {get; set;}
+    public double Calificacion {get; set;}
 
 }
